+++ app/src/main/java/com/example/pomodoro/MainActivity.java
@@ -15,7 +15,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.example.pomodoro.SetTime.SettimeActivity;
+import com.example.pomodoro.SetTimeFunctions.SettimeActivity;
 import com.example.pomodoro.TasksFunctions.TasksActivity;
 
 import java.util.Locale;
