+++ app/src/main/res/drawable/custom_checkbox.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:drawable="@drawable/checked"
+        android:state_checked="true"/>
+
+    <item android:drawable="@drawable/unchecked"/>
+</selector>
